refactor(sykmelding): extract domain models and metrics to separate files
Move SykmeldingsperiodeToStore and PersistedSykmeldingsperiode to
db/domain/ package and extract metric counters to SykmeldingsperiodeMetrics.kt,
following existing repo patterns (OppfolginsplanAPiMetrics, db/domain/).
