+++ src/main/kotlin/no/nav/syfo/sykmelding/kafka/SykmeldingsperiodeConsumer.kt
@@ -2,7 +2,6 @@ package no.nav.syfo.sykmelding.kafka
 
 import com.fasterxml.jackson.core.JsonProcessingException
 import com.fasterxml.jackson.module.kotlin.readValue
-import io.micrometer.core.instrument.Counter
 import kotlinx.coroutines.CancellationException
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.currentCoroutineContext
@@ -11,10 +10,8 @@ import kotlinx.coroutines.isActive
 import kotlinx.coroutines.withContext
 import no.nav.syfo.application.kafka.KafkaEnv
 import no.nav.syfo.application.kafka.consumerProperties
-import no.nav.syfo.application.metric.METRICS_NS
-import no.nav.syfo.application.metric.METRICS_REGISTRY
 import no.nav.syfo.sykmelding.db.SykmeldingsperiodeRepository
-import no.nav.syfo.sykmelding.db.SykmeldingsperiodeToStore
+import no.nav.syfo.sykmelding.db.domain.SykmeldingsperiodeToStore
 import no.nav.syfo.sykmelding.kafka.model.SendtSykmeldingKafkaMessage
 import no.nav.syfo.util.configuredJacksonMapper
 import no.nav.syfo.util.logger
@@ -28,19 +25,6 @@ import kotlin.time.Duration.Companion.seconds
 
 const val SYKMELDINGSPERIODE_TOPIC = "teamsykmelding.syfo-sendt-sykmelding"
 const val SYKMELDINGSPERIODE_CONSUMER_GROUP = "syfo-oppfolgingsplan-backend-sykmeldingsperiode"
-
-val COUNT_SYKMELDING_CONSUMED: Counter = Counter.builder("${METRICS_NS}_sykmelding_consumed")
-    .description("Counts the number of sykmeldingsperioder stored from Kafka")
-    .register(METRICS_REGISTRY)
-val COUNT_SYKMELDING_TOMBSTONE: Counter = Counter.builder("${METRICS_NS}_sykmelding_tombstone")
-    .description("Counts the number of sykmelding tombstones processed from Kafka")
-    .register(METRICS_REGISTRY)
-val COUNT_SYKMELDING_DESERIALIZATION_ERROR: Counter = Counter.builder("${METRICS_NS}_sykmelding_deserialization_error")
-    .description("Counts the number of sykmelding Kafka messages that failed deserialization (poison pills)")
-    .register(METRICS_REGISTRY)
-val COUNT_SYKMELDING_RUNTIME_ERROR: Counter = Counter.builder("${METRICS_NS}_sykmelding_runtime_error")
-    .description("Counts transient consumer runtime errors (connection, commit, etc.)")
-    .register(METRICS_REGISTRY)
 
 class SykmeldingsperiodeConsumer(
     private val sykmeldingsperiodeRepository: SykmeldingsperiodeRepository,
